projects/fraud_detection_system/requirements.txt
# Core ML/Data
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
imbalanced-learn==0.11.0

# Kafka
kafka-python==2.0.2

# Redis
redis==5.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# API
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3

# Monitoring/Visualization
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# MLflow
mlflow==2.9.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
faker==22.0.0
